+++ Mods ChoGGi/Random Starting Sector/Code/Script.lua
@@ -32,6 +32,7 @@ function Colony:CreateMapRand(which, ...)
 	return ChoOrig_Colony_CreateMapRand(self, which, ...)
 end
 
+-- I know I disabled this for a reason, but what?
 --~ local ChoOrig_InitialReveal = InitialReveal
 function InitialReveal(eligible, ...)
 --~ 	if mod_MinimumSurfaceDeposits == 0 and mod_MinimumSubsurfaceDeposits == 0 then