+++ Mods ChoGGi/Restore UI BBB/metadata.lua
@@ -4,9 +4,9 @@ return PlaceObj("ModDef", {
 	"steam_id", "2710662261",
 	"pops_any_uuid", "36625d3b-b3d4-4eef-af36-17797d7e2f0e",
 	"lua_revision", 1007000, -- Picard
-	"version", 1,
+	"version", 2,
 	"version_major", 0,
-	"version_minor", 1,
+	"version_minor", 2,
 	"image", "Preview.jpg",
 	"author", "ChoGGi",
 	"code", {
@@ -16,7 +16,6 @@ return PlaceObj("ModDef", {
 	"TagInterface", true,
 	"description", [[
 BBB == Before Below & Beyond
-
 Restores missing selection panel info:
 
 Max drone count for drone hubs.